1. 스킬 작업 2. 카메라 작업

+++ ARPG/Assets/Scripts/Data/AutoCreateExcelData/Common/Data.Content.ConstValue.cs
@@ -41,7 +41,7 @@ namespace Data.Contents
     public class ConstValueScript
     {
 		public Define.ConstDefType ConstType { get; set; }
-		public float Value { get; set; }
+		public int Value { get; set; }
 
     }
 }
